requirements.txt
docopt===0.6.2
num2words===0.5.14
django_widget_tweaks===1.5.0
pillow===11.3.0
fonttools===4.59.0
defusedxml===0.7.1
fpdf2===2.8.3